src/knowledge_graph/distributed_knowledge_graph.py
import asyncio
import ray
import numpy as np
import networkx as nx
from typing import Dict, List, Any, Tuple
from concurrent.futures import ProcessPoolExecutor
from dataclasses import dataclass, field

@dataclass
class DistributedKnowledgeNode:
    """
    Distributed Knowledge Graph Node with quantum-inspired properties
    """
    id: str
    data: Dict[str, Any]
    quantum_state: np.ndarray = field(default_factory=lambda: np.random.rand(10))
    entanglement_score: float = 0.0
    coherence_factor: float = 1.0

class DistributedKnowledgeGraphEngine:
    def __init__(self, 
                 num_workers: int = 4, 
                 partition_strategy: str = 'semantic'):
        """
        Initialize Distributed Knowledge Graph Engine
        
        Args:
            num_workers: Number of distributed processing workers
            partition_strategy: Strategy for graph partitioning
        """
        # Initialize Ray for distributed computing
        ray.init(num_cpus=num_workers)
        
        # Graph partitioning parameters
        self.num_workers = num_workers
        self.partition_strategy = partition_strategy
        
        # Quantum-inspired graph representation
        self.quantum_graph = nx.DiGraph()
        
        # Distributed processing executor
        self.executor = ProcessPoolExecutor(max_workers=num_workers)
    
    @ray.remote
    def _quantum_state_computation(self, node_data: Dict[str, Any]) -> np.ndarray:
        """
        Compute quantum-inspired state for a node
        
        Args:
            node_data: Node data for state computation
        
        Returns:
            Quantum-inspired state vector
        """
        # Quantum state generation using probabilistic encoding
        base_embedding = np.random.rand(10)  # Initial random state
        
        # Encode node metadata into quantum state
        for key, value in node_data.items():
            # Use hash-based encoding to introduce deterministic randomness
            hash_value = hash(str(value)) % 1000 / 1000.0
            base_embedding += np.sin(hash_value * np.pi) * base_embedding
        
        # Normalize to maintain quantum state properties
        return base_embedding / np.linalg.norm(base_embedding)
    
    def semantic_graph_partitioning(self, knowledge_graph: nx.DiGraph) -> List[nx.DiGraph]:
        """
        Partition knowledge graph using semantic clustering
        
        Args:
            knowledge_graph: Input knowledge graph
        
        Returns:
            List of semantic subgraphs
        """
        from sklearn.cluster import SpectralCoclustering
        
        # Compute adjacency matrix
        adjacency_matrix = nx.adjacency_matrix(knowledge_graph).todense()
        
        # Perform spectral co-clustering
        n_clusters = self.num_workers
        model = SpectralCoclustering(n_clusters=n_clusters)
        model.fit(adjacency_matrix)
        
        # Create subgraphs based on clustering
        partitioned_graphs = [nx.DiGraph() for _ in range(n_clusters)]
        
        for node_idx, cluster_label in enumerate(model.row_labels_):
            node = list(knowledge_graph.nodes())[node_idx]
            partitioned_graphs[cluster_label].add_node(node, **knowledge_graph.nodes[node])
        
        return partitioned_graphs
    
    async def distributed_graph_computation(self, 
                                            knowledge_graph: nx.DiGraph, 
                                            computation_func: callable):
        """
        Distribute graph computation across multiple workers
        
        Args:
            knowledge_graph: Input knowledge graph
            computation_func: Function to apply on graph partitions
        
        Returns:
            Aggregated computation results
        """
        # Partition graph
        graph_partitions = self.semantic_graph_partitioning(knowledge_graph)
        
        # Distribute computation
        computation_futures = [
            ray.remote(computation_func).remote(partition) 
            for partition in graph_partitions
        ]
        
        # Wait and aggregate results
        results = await ray.get(computation_futures)
        
        return results
    
    def quantum_entanglement_analysis(self) -> Dict[str, float]:
        """
        Analyze quantum entanglement properties of knowledge graph
        
        Returns:
            Entanglement scores for nodes
        """
        entanglement_scores = {}
        
        for node in self.quantum_graph.nodes():
            # Compute local entanglement based on neighborhood
            neighbors = list(self.quantum_graph.neighbors(node))
            
            if neighbors:
                # Compute quantum-inspired entanglement
                neighbor_states = [
                    self.quantum_graph.nodes[neighbor]['quantum_state'] 
                    for neighbor in neighbors
                ]
                
                # Compute correlation tensor
                correlation_tensor = np.mean(neighbor_states, axis=0)
                
                # Entanglement score based on state correlation
                entanglement_score = np.linalg.norm(correlation_tensor)
                
                entanglement_scores[node] = entanglement_score
        
        return entanglement_scores
    
    def add_quantum_node(self, node_id: str, node_data: Dict[str, Any]):
        """
        Add a node with quantum-inspired properties
        
        Args:
            node_id: Unique node identifier
            node_data: Node metadata
        """
        # Compute quantum state
        quantum_state = ray.get(
            self._quantum_state_computation.remote(node_data)
        )
        
        # Add node to quantum graph
        self.quantum_graph.add_node(node_id, 
                                    data=node_data, 
                                    quantum_state=quantum_state)
    
    def quantum_graph_inference(self, 
                                query_node: str, 
                                inference_depth: int = 3) -> List[Tuple[str, float]]:
        """
        Perform quantum-inspired graph inference
        
        Args:
            query_node: Starting node for inference
            inference_depth: Depth of inference traversal
        
        Returns:
            List of inferred nodes with relevance scores
        """
        # Quantum walk inference
        inferred_nodes = []
        visited = set()
        
        def quantum_walk(current_node, depth):
            if depth == 0 or current_node in visited:
                return
            
            visited.add(current_node)
            current_state = self.quantum_graph.nodes[current_node]['quantum_state']
            
            # Explore neighbors
            for neighbor in self.quantum_graph.neighbors(current_node):
                neighbor_state = self.quantum_graph.nodes[neighbor]['quantum_state']
                
                # Compute quantum-inspired relevance
                relevance_score = np.dot(current_state, neighbor_state)
                
                inferred_nodes.append((neighbor, relevance_score))
                
                # Recursive quantum walk
                quantum_walk(neighbor, depth - 1)
        
        # Start quantum walk
        quantum_walk(query_node, inference_depth)
        
        # Sort by relevance score
        return sorted(inferred_nodes, key=lambda x: x[1], reverse=True)

# Cleanup function for distributed resources
def cleanup_distributed_resources():
    ray.shutdown()
